backend/requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openai
python-dateutil
pytz
nltk
sumy
pandas
